requirements.txt
matplotlib 
scikit-learn 
seaborn 
pandas 
numpy 
werkzeug
plotly.express 
joblib
# io
tensorflow 


